Reorganized GraphQL Schema files
