+++ schema/index.ts
@@ -1,0 +1,2 @@
+export { typeDefs } from './type-defs';
+export { resolvers } from './resolvers';
